[models] move TransferFunction to separate module
